+++ src/profiler.cpp
@@ -1,1 +1,3 @@
-unsigned long long average_loop_micros = 0;
+#include "Arduino.h"
+
+uint32_t average_loop_micros = 0;
